tools: fix int32 typedef

+++ tools/decode_map/util.h
@@ -10,8 +10,8 @@ typedef unsigned char uint8;
 typedef signed char int8;
 typedef unsigned short uint16;
 typedef signed short int16;
-typedef unsigned long uint32;
-typedef signed long int32;
+typedef unsigned int uint32;
+typedef signed int int32;
 
 extern uint16 read_uint16LE(const void *ptr);
 extern uint32 read_uint32LE(const void *ptr);
